backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
cloudinary
python-jose
supabase
pydantic
psycopg2-binary
